restucturing L2 - renaming folder name/path
