+++ src/configs/properties.js
