+++ src/models/tips.js
