+++ src/resources/docker/keycloak.yml
@@ -1,5 +1,5 @@
 # This configuration is intended for development purpose, it's **your** responsibility to harden it for production
-name: wda-kc
+name: wedaa-kc
 services:
     keycloak:
         image: quay.io/keycloak/keycloak:20.0.1
