+++ src/resources/docker/mongodb.yml
@@ -1,7 +1,7 @@
 # This configuration is intended for development purpose, it's **your** responsibility to harden it for production
 version: '3.8'
 services:
-    wda-mongodb:
+    wedaa-mongodb:
         image: mongo:4.4.15
         # If you want to expose these ports outside your dev PC,
         # remove the "127.0.0.1:" prefix
